+++ src/gemini-api-client/types.ts
@@ -1,6 +1,7 @@
 import type { components } from "../generated-types/gemini-types.ts"
 
 export type GenerateContentRequest = components["schemas"]["GenerateContentRequest"]
+
 export type Content = components["schemas"]["Content"]
 
 export type GenerateContentResponse = components["schemas"]["GenerateContentResponse"]
@@ -14,20 +15,9 @@ export type FinishReason = Candidate["finishReason"]
 export type Part = components["schemas"]["Part"]
 
 export type EmbedContentRequest = components["schemas"]["EmbedContentRequest"]
+
 export type EmbedContentResponse = components["schemas"]["EmbedContentResponse"]
 
 export interface RequestOptions {
   timeout: number
-}
-
-export interface EnhancedGenerateContentResponse extends GenerateContentResponse {
-  /**
-   * Returns the text string from the response, if available.
-   * Throws if the prompt or candidate was blocked.
-   */
-  result: () => string | FunctionCall
-}
-
-export interface GenerateContentResult {
-  response: EnhancedGenerateContentResponse
 }
